revert 6630fe7cc0b5a393a54e6c6df20682685a91fb4c availableForWrite() return int

+++ cores/arduino/USB/USBAPI.h
@@ -129,7 +129,7 @@ public:
 	void end(void);
 
 	virtual int available(void);
-	virtual size_t availableForWrite(void);
+	virtual int availableForWrite(void);
 	virtual int peek(void);
 	virtual int read(void);
 	virtual void flush(void);
